fix(storybook): pin the foundations pages to variables that exist
Four rows read `--secondary-hover`, `--secondary-active`, `--ghost-active` and
`--destructive-hover`. Those compatibility tokens never left the `--cs-`
namespace — they reach components only through `--color-*` — so the swatches
resolved to nothing. Uniwind logs one dev warning for the first miss in a
session and stays quiet afterwards, so three of the four were invisible.

The weight table read `--font-weight-*`, which lives in `@theme inline static`:
Tailwind substitutes the value into the utility, so the variable need not exist
at runtime at all. Reading it was a bet. The rows now carry the claim instead —
`font-bold` is remapped to 600, so it must render identically to
`font-semibold`, which is a stronger check than a number anyway.

Splits the name tables into `tokens.ts` so a test can walk them without pulling
in a React tree, and asserts each name against the CSS the build emits —
counting plain `:root`/`@variant` declarations and `@theme static`, but not the
two blocks Tailwind may inline or prune. Verified the guard is not vacuous by
reintroducing the original typo, which fails it.
