+++ packages/ui/stories/foundations/materials.stories.tsx
@@ -3,41 +3,14 @@ import { Text, View } from 'react-native';
 import { useCSSVariable } from 'uniwind';
 
 import { FoundationPage, formatTokenValue, Group, SpecRow, ThemeSplit } from './showcase';
-
-/**
- * `--radius` is the single authored step (VBG ships only 6px and 8px); every
- * `rounded-sm…4xl` utility is a multiple of it, generated by
- * `packages/design-tokens/scripts/build-native-css.ts`. The derivation is shown
- * instead of a baked px so the table cannot drift away from that script.
- */
-const RADIUS_STEPS = [
-  { className: 'rounded-4xs', derivation: '0.5px' },
-  { className: 'rounded-3xs', derivation: '1px' },
-  { className: 'rounded-2xs', derivation: '1.5px' },
-  { className: 'rounded-xs', derivation: '2px' },
-  { className: 'rounded-sm', derivation: '--radius × 0.6' },
-  { className: 'rounded-md', derivation: '--radius × 0.8' },
-  { className: 'rounded-lg', derivation: '--radius' },
-  { className: 'rounded-xl', derivation: '--radius × 1.4' },
-  { className: 'rounded-2xl', derivation: '--radius × 1.8' },
-  { className: 'rounded-3xl', derivation: '--radius × 2.2' },
-  { className: 'rounded-4xl', derivation: '--radius × 2.6' },
-  { className: 'rounded-round', derivation: '9999px' },
-];
-
-const BORDER_RAMP = [
-  { className: 'bg-border-subtle', name: '--border-subtle', usage: '同一表面内的分隔线' },
-  { className: 'bg-border', name: '--border', usage: '默认描边' },
-  { className: 'bg-border-strong', name: '--border-strong', usage: '需要被看见的分组边界' },
-  { className: 'bg-border-selected', name: '--border-selected', usage: '选中态' },
-];
+import { BORDER_RAMP, RADIUS_STEPS, RADIUS_VARIABLE } from './tokens';
 
 function RadiusRows() {
-  const radius = useCSSVariable('--radius');
+  const radius = useCSSVariable(RADIUS_VARIABLE);
 
   return (
     <View className="gap-4">
-      <SpecRow name="--radius" value={formatTokenValue(radius)} />
+      <SpecRow name={RADIUS_VARIABLE} value={formatTokenValue(radius)} />
       <View className="gap-3">
         {RADIUS_STEPS.map(({ className, derivation }) => (
           <View className="flex-row items-center gap-3" key={className}>
